+++ c/examples/api_structure.c
@@ -17,7 +17,7 @@ main(int argc, char **argv)
     ret = tsk_edge_table_init(&edges, 0);
     check_tsk_error(ret);
     for (j = 0; j < 5; j++) {
-        ret = tsk_edge_table_add_row(&edges, 0, 1, j + 1, j);
+        ret = tsk_edge_table_add_row(&edges, 0, 1, j + 1, j, NULL, 0);
         check_tsk_error(ret);
     }
     tsk_edge_table_print_state(&edges, stdout);