+++ c/tskit/tables.c
@@ -42,12 +42,15 @@
 #define DEFAULT_SIZE_INCREMENT 1024
 #define TABLE_SEP "-----------------------------------------\n"
 
+#define TSK_COL_OPTIONAL (1 << 0)
+
 typedef struct {
     const char *name;
     void **array_dest;
     tsk_size_t *len_dest;
     tsk_size_t len_offset;
     int type;
+    tsk_flags_t options;
 } read_table_col_t;
 
 typedef struct {
@@ -94,22 +97,32 @@ read_table_cols(kastore_t *store, read_table_col_t *read_cols, size_t num_cols)
         *read_cols[j].len_dest = (tsk_size_t) -1;
     }
     for (j = 0; j < num_cols; j++) {
-        ret = kastore_gets(
-            store, read_cols[j].name, read_cols[j].array_dest, &len, &type);
-        if (ret != 0) {
-            ret = tsk_set_kas_error(ret);
+        if (kastore_containss(store, read_cols[j].name)) {
+            ret = kastore_gets(
+                store, read_cols[j].name, read_cols[j].array_dest, &len, &type);
+            if (ret != 0) {
+                ret = tsk_set_kas_error(ret);
+                goto out;
+            }
+            last_len = *read_cols[j].len_dest;
+            if (last_len == (tsk_size_t) -1) {
+                *read_cols[j].len_dest = (tsk_size_t)(len - read_cols[j].len_offset);
+            } else if ((last_len + read_cols[j].len_offset) != (tsk_size_t) len) {
+                ret = TSK_ERR_FILE_FORMAT;
+                goto out;
+            }
+            if (type != read_cols[j].type) {
+                ret = TSK_ERR_FILE_FORMAT;
+                goto out;
+            }
+        } else if (!(read_cols[j].options & TSK_COL_OPTIONAL)) {
+            ret = TSK_REQUIRED_COL_NOT_FOUND;
             goto out;
-        }
-        last_len = *read_cols[j].len_dest;
-        if (last_len == (tsk_size_t) -1) {
-            *read_cols[j].len_dest = (tsk_size_t)(len - read_cols[j].len_offset);
-        } else if ((last_len + read_cols[j].len_offset) != (tsk_size_t) len) {
-            ret = TSK_ERR_FILE_FORMAT;
-            goto out;
-        }
-        if (type != read_cols[j].type) {
-            ret = TSK_ERR_FILE_FORMAT;
-            goto out;
+        } else {
+            read_cols[j].array_dest = NULL;
+            if (*read_cols[j].len_dest == (tsk_size_t) -1) {
+                *read_cols[j].len_dest = 0;
+            }
         }
     }
 out:
@@ -664,15 +677,16 @@ static int
 tsk_individual_table_load(tsk_individual_table_t *self, kastore_t *store)
 {
     read_table_col_t read_cols[] = {
-        { "individuals/flags", (void **) &self->flags, &self->num_rows, 0, KAS_UINT32 },
+        { "individuals/flags", (void **) &self->flags, &self->num_rows, 0, KAS_UINT32,
+            0 },
         { "individuals/location", (void **) &self->location, &self->location_length, 0,
-            KAS_FLOAT64 },
+            KAS_FLOAT64, 0 },
         { "individuals/location_offset", (void **) &self->location_offset,
-            &self->num_rows, 1, KAS_UINT32 },
+            &self->num_rows, 1, KAS_UINT32, 0 },
         { "individuals/metadata", (void **) &self->metadata, &self->metadata_length, 0,
-            KAS_UINT8 },
+            KAS_UINT8, 0 },
         { "individuals/metadata_offset", (void **) &self->metadata_offset,
-            &self->num_rows, 1, KAS_UINT32 },
+            &self->num_rows, 1, KAS_UINT32, 0 },
     };
     return read_table_cols(store, read_cols, sizeof(read_cols) / sizeof(*read_cols));
 }
@@ -1081,16 +1095,16 @@ static int
 tsk_node_table_load(tsk_node_table_t *self, kastore_t *store)
 {
     read_table_col_t read_cols[] = {
-        { "nodes/time", (void **) &self->time, &self->num_rows, 0, KAS_FLOAT64 },
-        { "nodes/flags", (void **) &self->flags, &self->num_rows, 0, KAS_UINT32 },
-        { "nodes/population", (void **) &self->population, &self->num_rows, 0,
-            KAS_INT32 },
-        { "nodes/individual", (void **) &self->individual, &self->num_rows, 0,
-            KAS_INT32 },
+        { "nodes/time", (void **) &self->time, &self->num_rows, 0, KAS_FLOAT64, 0 },
+        { "nodes/flags", (void **) &self->flags, &self->num_rows, 0, KAS_UINT32, 0 },
+        { "nodes/population", (void **) &self->population, &self->num_rows, 0, KAS_INT32,
+            0 },
+        { "nodes/individual", (void **) &self->individual, &self->num_rows, 0, KAS_INT32,
+            0 },
         { "nodes/metadata", (void **) &self->metadata, &self->metadata_length, 0,
-            KAS_UINT8 },
+            KAS_UINT8, 0 },
         { "nodes/metadata_offset", (void **) &self->metadata_offset, &self->num_rows, 1,
-            KAS_UINT32 },
+            KAS_UINT32, 0 },
     };
     return read_table_cols(store, read_cols, sizeof(read_cols) / sizeof(*read_cols));
 }
@@ -1100,7 +1114,32 @@ tsk_node_table_load(tsk_node_table_t *self, kastore_t *store)
  *************************/
 
 static int
-tsk_edge_table_expand_columns(tsk_edge_table_t *self, size_t additional_rows)
+tsk_edge_table_metadata_init(tsk_edge_table_t *self)
+{
+    int ret = 0;
+
+    if (self->metadata == NULL) {
+        ret = expand_column(
+            (void **) &self->metadata_offset, self->max_rows + 1, sizeof(tsk_size_t));
+        if (ret != 0) {
+            goto out;
+        }
+        memset(self->metadata_offset, 0, (self->max_rows + 1) * sizeof(tsk_size_t));
+
+        ret = expand_column((void **) &self->metadata, 1, sizeof(char));
+        if (ret != 0) {
+            goto out;
+        }
+        self->max_metadata_length = 1;
+        self->metadata_malloced_locally = true;
+    }
+
+out:
+    return ret;
+}
+
+static int
+tsk_edge_table_expand_main_columns(tsk_edge_table_t *self, size_t additional_rows)
 {
     int ret = 0;
     tsk_size_t increment
@@ -1128,7 +1167,35 @@ tsk_edge_table_expand_columns(tsk_edge_table_t *self, size_t additional_rows)
         if (ret != 0) {
             goto out;
         }
+        ret = expand_column(
+            (void **) &self->metadata_offset, new_size + 1, sizeof(tsk_size_t));
+        if (ret != 0) {
+            goto out;
+        }
         self->max_rows = new_size;
+    }
+out:
+    return ret;
+}
+
+static int
+tsk_edge_table_expand_metadata(tsk_edge_table_t *self, tsk_size_t additional_length)
+{
+    int ret = 0;
+    tsk_size_t increment
+        = TSK_MAX(additional_length, self->max_metadata_length_increment);
+    tsk_size_t new_size = self->max_metadata_length + increment;
+
+    if (check_offset_overflow(self->metadata_length, increment)) {
+        ret = TSK_ERR_COLUMN_OVERFLOW;
+        goto out;
+    }
+    if ((self->metadata_length + additional_length) > self->max_metadata_length) {
+        ret = expand_column((void **) &self->metadata, new_size, sizeof(char));
+        if (ret != 0) {
+            goto out;
+        }
+        self->max_metadata_length = new_size;
     }
 out:
     return ret;
@@ -1146,6 +1213,17 @@ tsk_edge_table_set_max_rows_increment(
 }
 
 int
+tsk_edge_table_set_max_metadata_length_increment(
+    tsk_edge_table_t *self, tsk_size_t max_metadata_length_increment)
+{
+    if (max_metadata_length_increment == 0) {
+        max_metadata_length_increment = DEFAULT_SIZE_INCREMENT;
+    }
+    self->max_metadata_length_increment = max_metadata_length_increment;
+    return 0;
+}
+
+int
 tsk_edge_table_init(tsk_edge_table_t *self, tsk_flags_t TSK_UNUSED(options))
 {
     int ret = 0;
@@ -1155,29 +1233,48 @@ tsk_edge_table_init(tsk_edge_table_t *self, tsk_flags_t TSK_UNUSED(options))
     /* Allocate space for one row initially, ensuring we always have valid pointers
      * even if the table is empty */
     self->max_rows_increment = 1;
-    ret = tsk_edge_table_expand_columns(self, 1);
+    self->max_metadata_length_increment = 1;
+    ret = tsk_edge_table_expand_main_columns(self, 1);
     if (ret != 0) {
         goto out;
     }
+    ret = tsk_edge_table_expand_metadata(self, 1);
+    if (ret != 0) {
+        goto out;
+    }
+    self->metadata_offset[0] = 0;
+    self->metadata_malloced_locally = true;
     self->max_rows_increment = DEFAULT_SIZE_INCREMENT;
+    self->max_metadata_length_increment = DEFAULT_SIZE_INCREMENT;
 out:
     return ret;
 }
 
 tsk_id_t
-tsk_edge_table_add_row(
-    tsk_edge_table_t *self, double left, double right, tsk_id_t parent, tsk_id_t child)
+tsk_edge_table_add_row(tsk_edge_table_t *self, double left, double right,
+    tsk_id_t parent, tsk_id_t child, const char *metadata, tsk_size_t metadata_length)
 {
     int ret = 0;
 
-    ret = tsk_edge_table_expand_columns(self, 1);
+    ret = tsk_edge_table_expand_main_columns(self, 1);
     if (ret != 0) {
         goto out;
     }
+    ret = tsk_edge_table_expand_metadata(self, metadata_length);
+    if (ret != 0) {
+        goto out;
+    }
+
+    assert(self->num_rows < self->max_rows);
+    assert(self->metadata_length + metadata_length <= self->max_metadata_length);
+    memcpy(self->metadata + self->metadata_length, metadata, metadata_length);
     self->left[self->num_rows] = left;
     self->right[self->num_rows] = right;
     self->parent[self->num_rows] = parent;
     self->child[self->num_rows] = child;
+    self->metadata_offset[self->num_rows + 1] = self->metadata_length + metadata_length;
+    self->metadata_length += metadata_length;
+
     ret = (tsk_id_t) self->num_rows;
     self->num_rows++;
 out:
@@ -1195,15 +1292,16 @@ tsk_edge_table_copy(tsk_edge_table_t *self, tsk_edge_table_t *dest, tsk_flags_t 
             goto out;
         }
     }
-    ret = tsk_edge_table_set_columns(
-        dest, self->num_rows, self->left, self->right, self->parent, self->child);
+    ret = tsk_edge_table_set_columns(dest, self->num_rows, self->left, self->right,
+        self->parent, self->child, self->metadata, self->metadata_offset);
 out:
     return ret;
 }
 
 int
 tsk_edge_table_set_columns(tsk_edge_table_t *self, tsk_size_t num_rows, double *left,
-    double *right, tsk_id_t *parent, tsk_id_t *child)
+    double *right, tsk_id_t *parent, tsk_id_t *child, const char *metadata,
+    tsk_size_t *metadata_offset)
 {
     int ret = 0;
 
@@ -1211,22 +1309,30 @@ tsk_edge_table_set_columns(tsk_edge_table_t *self, tsk_size_t num_rows, double *
     if (ret != 0) {
         goto out;
     }
-    ret = tsk_edge_table_append_columns(self, num_rows, left, right, parent, child);
+    ret = tsk_edge_table_append_columns(
+        self, num_rows, left, right, parent, child, metadata, metadata_offset);
 out:
     return ret;
 }
 
 int
 tsk_edge_table_append_columns(tsk_edge_table_t *self, tsk_size_t num_rows, double *left,
-    double *right, tsk_id_t *parent, tsk_id_t *child)
+    double *right, tsk_id_t *parent, tsk_id_t *child, const char *metadata,
+    tsk_size_t *metadata_offset)
 {
     int ret;
+    tsk_size_t j, metadata_length;
 
     if (left == NULL || right == NULL || parent == NULL || child == NULL) {
         ret = TSK_ERR_BAD_PARAM_VALUE;
         goto out;
     }
-    ret = tsk_edge_table_expand_columns(self, num_rows);
+    if ((metadata == NULL) != (metadata_offset == NULL)) {
+        ret = TSK_ERR_BAD_PARAM_VALUE;
+        goto out;
+    }
+
+    ret = tsk_edge_table_expand_main_columns(self, num_rows);
     if (ret != 0) {
         goto out;
     }
@@ -1234,7 +1340,30 @@ tsk_edge_table_append_columns(tsk_edge_table_t *self, tsk_size_t num_rows, doubl
     memcpy(self->right + self->num_rows, right, num_rows * sizeof(double));
     memcpy(self->parent + self->num_rows, parent, num_rows * sizeof(tsk_id_t));
     memcpy(self->child + self->num_rows, child, num_rows * sizeof(tsk_id_t));
+    if (metadata == NULL) {
+        for (j = 0; j < num_rows; j++) {
+            self->metadata_offset[self->num_rows + j + 1] = self->metadata_length;
+        }
+    } else {
+        ret = check_offsets(num_rows, metadata_offset, 0, false);
+        if (ret != 0) {
+            goto out;
+        }
+        for (j = 0; j < num_rows; j++) {
+            self->metadata_offset[self->num_rows + j]
+                = (tsk_size_t) self->metadata_length + metadata_offset[j];
+        }
+        metadata_length = metadata_offset[num_rows];
+        ret = tsk_edge_table_expand_metadata(self, metadata_length);
+        if (ret != 0) {
+            goto out;
+        }
+        memcpy(self->metadata + self->metadata_length, metadata,
+            metadata_length * sizeof(char));
+        self->metadata_length += metadata_length;
+    }
     self->num_rows += num_rows;
+    self->metadata_offset[self->num_rows] = self->metadata_length;
 out:
     return ret;
 }
@@ -1255,6 +1384,7 @@ tsk_edge_table_truncate(tsk_edge_table_t *self, tsk_size_t num_rows)
         goto out;
     }
     self->num_rows = num_rows;
+    self->metadata_length = self->metadata_offset[num_rows];
 out:
     return ret;
 }
@@ -1268,6 +1398,10 @@ tsk_edge_table_free(tsk_edge_table_t *self)
         tsk_safe_free(self->parent);
         tsk_safe_free(self->child);
     }
+    if (self->metadata_malloced_locally) {
+        tsk_safe_free(self->metadata);
+        tsk_safe_free(self->metadata_offset);
+    }
     return 0;
 }
 
@@ -1280,6 +1414,9 @@ tsk_edge_table_print_state(tsk_edge_table_t *self, FILE *out)
     fprintf(out, "edge_table: %p:\n", (void *) self);
     fprintf(out, "num_rows          = %d\tmax= %d\tincrement = %d)\n",
         (int) self->num_rows, (int) self->max_rows, (int) self->max_rows_increment);
+    fprintf(out, "metadata_length = %d\tmax= %d\tincrement = %d)\n",
+        (int) self->metadata_length, (int) self->max_metadata_length,
+        (int) self->max_metadata_length_increment);
     fprintf(out, TABLE_SEP);
     ret = tsk_edge_table_dump_text(self, out);
     assert(ret == 0);
@@ -1290,15 +1427,18 @@ tsk_edge_table_dump_text(tsk_edge_table_t *self, FILE *out)
 {
     tsk_size_t j;
     int ret = TSK_ERR_IO;
+    tsk_size_t metadata_len;
     int err;
 
-    err = fprintf(out, "id\tleft\tright\tparent\tchild\n");
+    err = fprintf(out, "id\tleft\tright\tparent\tchild\tmetadata\n");
     if (err < 0) {
         goto out;
     }
     for (j = 0; j < self->num_rows; j++) {
-        err = fprintf(out, "%d\t%.3f\t%.3f\t%d\t%d\n", j, self->left[j], self->right[j],
-            self->parent[j], self->child[j]);
+        metadata_len = self->metadata_offset[j + 1] - self->metadata_offset[j];
+        err = fprintf(out, "%d\t%.3f\t%.3f\t%d\t%d\t%.*s\n", j, self->left[j],
+            self->right[j], self->parent[j], self->child[j], metadata_len,
+            self->metadata + self->metadata_offset[j]);
         if (err < 0) {
             goto out;
         }
@@ -1312,12 +1452,19 @@ bool
 tsk_edge_table_equals(tsk_edge_table_t *self, tsk_edge_table_t *other)
 {
     bool ret = false;
-    if (self->num_rows == other->num_rows) {
+    if (self->num_rows == other->num_rows
+        && self->metadata_length == other->metadata_length) {
         ret = memcmp(self->left, other->left, self->num_rows * sizeof(double)) == 0
               && memcmp(self->right, other->right, self->num_rows * sizeof(double)) == 0
               && memcmp(self->parent, other->parent, self->num_rows * sizeof(tsk_id_t))
                      == 0
               && memcmp(self->child, other->child, self->num_rows * sizeof(tsk_id_t))
+                     == 0
+              && memcmp(self->metadata_offset, other->metadata_offset,
+                     (self->num_rows + 1) * sizeof(tsk_size_t))
+                     == 0
+              && memcmp(self->metadata, other->metadata,
+                     self->metadata_length * sizeof(char))
                      == 0;
     }
     return ret;
@@ -1336,6 +1483,9 @@ tsk_edge_table_get_row(tsk_edge_table_t *self, tsk_id_t index, tsk_edge_t *row)
     row->right = self->right[index];
     row->parent = self->parent[index];
     row->child = self->child[index];
+    row->metadata_length
+        = self->metadata_offset[index + 1] - self->metadata_offset[index];
+    row->metadata = self->metadata + self->metadata_offset[index];
 out:
     return ret;
 }
@@ -1348,6 +1498,9 @@ tsk_edge_table_dump(tsk_edge_table_t *self, kastore_t *store)
         { "edges/right", (void *) self->right, self->num_rows, KAS_FLOAT64 },
         { "edges/parent", (void *) self->parent, self->num_rows, KAS_INT32 },
         { "edges/child", (void *) self->child, self->num_rows, KAS_INT32 },
+        { "edges/metadata", (void *) self->metadata, self->metadata_length, KAS_UINT8 },
+        { "edges/metadata_offset", (void *) self->metadata_offset, self->num_rows + 1,
+            KAS_UINT32 },
     };
     return write_table_cols(store, write_cols, sizeof(write_cols) / sizeof(*write_cols));
 }
@@ -1355,13 +1508,37 @@ tsk_edge_table_dump(tsk_edge_table_t *self, kastore_t *store)
 static int
 tsk_edge_table_load(tsk_edge_table_t *self, kastore_t *store)
 {
+    int ret = 0;
+
     read_table_col_t read_cols[] = {
-        { "edges/left", (void **) &self->left, &self->num_rows, 0, KAS_FLOAT64 },
-        { "edges/right", (void **) &self->right, &self->num_rows, 0, KAS_FLOAT64 },
-        { "edges/parent", (void **) &self->parent, &self->num_rows, 0, KAS_INT32 },
-        { "edges/child", (void **) &self->child, &self->num_rows, 0, KAS_INT32 },
+        { "edges/left", (void **) &self->left, &self->num_rows, 0, KAS_FLOAT64, 0 },
+        { "edges/right", (void **) &self->right, &self->num_rows, 0, KAS_FLOAT64, 0 },
+        { "edges/parent", (void **) &self->parent, &self->num_rows, 0, KAS_INT32, 0 },
+        { "edges/child", (void **) &self->child, &self->num_rows, 0, KAS_INT32, 0 },
+        { "edges/metadata", (void **) &self->metadata, &self->metadata_length, 0,
+            KAS_UINT8, TSK_COL_OPTIONAL },
+        { "edges/metadata_offset", (void **) &self->metadata_offset, &self->num_rows, 1,
+            KAS_UINT32, TSK_COL_OPTIONAL },
+
     };
-    return read_table_cols(store, read_cols, sizeof(read_cols) / sizeof(*read_cols));
+    ret = read_table_cols(store, read_cols, sizeof(read_cols) / sizeof(*read_cols));
+    if (ret != 0) {
+        goto out;
+    }
+    if ((self->metadata == NULL) != (self->metadata_offset == NULL)) {
+        ret = TSK_REQUIRED_COL_NOT_FOUND;
+        goto out;
+    }
+    if (self->metadata == NULL) {
+        tsk_edge_table_metadata_init(self);
+    } else {
+        // As tsk_edge_table_init may not have been run before this function we set
+        // this as it won't have been init'd
+        self->metadata_malloced_locally = false;
+    }
+
+out:
+    return ret;
 }
 
 int
@@ -1371,6 +1548,11 @@ tsk_edge_table_squash(tsk_edge_table_t *self)
     int ret = 0;
     tsk_edge_t *edges = NULL;
     tsk_size_t num_output_edges;
+
+    if (self->metadata_length > 0) {
+        ret = TSK_ERR_CANT_PROCESS_EDGES_WITH_METADATA;
+        goto out;
+    }
 
     edges = malloc(self->num_rows * sizeof(tsk_edge_t));
     if (edges == NULL) {
@@ -1383,6 +1565,7 @@ tsk_edge_table_squash(tsk_edge_table_t *self)
         edges[k].right = self->right[k];
         edges[k].parent = self->parent[k];
         edges[k].child = self->child[k];
+        edges[k].metadata_length = 0;
     }
 
     ret = tsk_squash_edges(edges, self->num_rows, &num_output_edges);
@@ -1859,15 +2042,16 @@ static int
 tsk_site_table_load(tsk_site_table_t *self, kastore_t *store)
 {
     read_table_col_t read_cols[] = {
-        { "sites/position", (void **) &self->position, &self->num_rows, 0, KAS_FLOAT64 },
+        { "sites/position", (void **) &self->position, &self->num_rows, 0, KAS_FLOAT64,
+            0 },
         { "sites/ancestral_state", (void **) &self->ancestral_state,
-            &self->ancestral_state_length, 0, KAS_UINT8 },
+            &self->ancestral_state_length, 0, KAS_UINT8, 0 },
         { "sites/ancestral_state_offset", (void **) &self->ancestral_state_offset,
-            &self->num_rows, 1, KAS_UINT32 },
+            &self->num_rows, 1, KAS_UINT32, 0 },
         { "sites/metadata", (void **) &self->metadata, &self->metadata_length, 0,
-            KAS_UINT8 },
+            KAS_UINT8, 0 },
         { "sites/metadata_offset", (void **) &self->metadata_offset, &self->num_rows, 1,
-            KAS_UINT32 },
+            KAS_UINT32, 0 },
     };
     return read_table_cols(store, read_cols, sizeof(read_cols) / sizeof(*read_cols));
 }
@@ -2353,17 +2537,18 @@ static int
 tsk_mutation_table_load(tsk_mutation_table_t *self, kastore_t *store)
 {
     read_table_col_t read_cols[] = {
-        { "mutations/site", (void **) &self->site, &self->num_rows, 0, KAS_INT32 },
-        { "mutations/node", (void **) &self->node, &self->num_rows, 0, KAS_INT32 },
-        { "mutations/parent", (void **) &self->parent, &self->num_rows, 0, KAS_INT32 },
+        { "mutations/site", (void **) &self->site, &self->num_rows, 0, KAS_INT32, 0 },
+        { "mutations/node", (void **) &self->node, &self->num_rows, 0, KAS_INT32, 0 },
+        { "mutations/parent", (void **) &self->parent, &self->num_rows, 0, KAS_INT32,
+            0 },
         { "mutations/derived_state", (void **) &self->derived_state,
-            &self->derived_state_length, 0, KAS_UINT8 },
+            &self->derived_state_length, 0, KAS_UINT8, 0 },
         { "mutations/derived_state_offset", (void **) &self->derived_state_offset,
-            &self->num_rows, 1, KAS_UINT32 },
+            &self->num_rows, 1, KAS_UINT32, 0 },
         { "mutations/metadata", (void **) &self->metadata, &self->metadata_length, 0,
-            KAS_UINT8 },
+            KAS_UINT8, 0 },
         { "mutations/metadata_offset", (void **) &self->metadata_offset, &self->num_rows,
-            1, KAS_UINT32 },
+            1, KAS_UINT32, 0 },
     };
     return read_table_cols(store, read_cols, sizeof(read_cols) / sizeof(*read_cols));
 }
@@ -2655,12 +2840,14 @@ static int
 tsk_migration_table_load(tsk_migration_table_t *self, kastore_t *store)
 {
     read_table_col_t read_cols[] = {
-        { "migrations/left", (void **) &self->left, &self->num_rows, 0, KAS_FLOAT64 },
-        { "migrations/right", (void **) &self->right, &self->num_rows, 0, KAS_FLOAT64 },
-        { "migrations/node", (void **) &self->node, &self->num_rows, 0, KAS_INT32 },
-        { "migrations/source", (void **) &self->source, &self->num_rows, 0, KAS_INT32 },
-        { "migrations/dest", (void **) &self->dest, &self->num_rows, 0, KAS_INT32 },
-        { "migrations/time", (void **) &self->time, &self->num_rows, 0, KAS_FLOAT64 },
+        { "migrations/left", (void **) &self->left, &self->num_rows, 0, KAS_FLOAT64, 0 },
+        { "migrations/right", (void **) &self->right, &self->num_rows, 0, KAS_FLOAT64,
+            0 },
+        { "migrations/node", (void **) &self->node, &self->num_rows, 0, KAS_INT32, 0 },
+        { "migrations/source", (void **) &self->source, &self->num_rows, 0, KAS_INT32,
+            0 },
+        { "migrations/dest", (void **) &self->dest, &self->num_rows, 0, KAS_INT32, 0 },
+        { "migrations/time", (void **) &self->time, &self->num_rows, 0, KAS_FLOAT64, 0 },
     };
     return read_table_cols(store, read_cols, sizeof(read_cols) / sizeof(*read_cols));
 }
@@ -3003,9 +3190,9 @@ tsk_population_table_load(tsk_population_table_t *self, kastore_t *store)
 {
     read_table_col_t read_cols[] = {
         { "populations/metadata", (void **) &self->metadata, &self->metadata_length, 0,
-            KAS_UINT8 },
+            KAS_UINT8, 0 },
         { "populations/metadata_offset", (void **) &self->metadata_offset,
-            &self->num_rows, 1, KAS_UINT32 },
+            &self->num_rows, 1, KAS_UINT32, 0 },
     };
     return read_table_cols(store, read_cols, sizeof(read_cols) / sizeof(*read_cols));
 }
@@ -3448,13 +3635,13 @@ tsk_provenance_table_load(tsk_provenance_table_t *self, kastore_t *store)
 {
     read_table_col_t read_cols[] = {
         { "provenances/timestamp", (void **) &self->timestamp, &self->timestamp_length,
-            0, KAS_UINT8 },
+            0, KAS_UINT8, 0 },
         { "provenances/timestamp_offset", (void **) &self->timestamp_offset,
-            &self->num_rows, 1, KAS_UINT32 },
+            &self->num_rows, 1, KAS_UINT32, 0 },
         { "provenances/record", (void **) &self->record, &self->record_length, 0,
-            KAS_UINT8 },
+            KAS_UINT8, 0 },
         { "provenances/record_offset", (void **) &self->record_offset, &self->num_rows,
-            1, KAS_UINT32 },
+            1, KAS_UINT32, 0 },
     };
     return read_table_cols(store, read_cols, sizeof(read_cols) / sizeof(*read_cols));
 }
@@ -4044,7 +4231,8 @@ ancestor_mapper_flush_edges(
         child = self->buffered_children[j];
         for (x = self->child_edge_map_head[child]; x != NULL; x = x->next) {
             // printf("Adding edge %f %f %i %i\n", x->left, x->right, parent, child);
-            ret = tsk_edge_table_add_row(self->result, x->left, x->right, parent, child);
+            ret = tsk_edge_table_add_row(
+                self->result, x->left, x->right, parent, child, NULL, 0);
             if (ret < 0) {
                 goto out;
             }
@@ -4701,7 +4889,7 @@ simplifier_flush_edges(simplifier_t *self, tsk_id_t parent, size_t *ret_num_edge
         child = self->buffered_children[j];
         for (x = self->child_edge_map_head[child]; x != NULL; x = x->next) {
             ret = tsk_edge_table_add_row(
-                &self->tables->edges, x->left, x->right, parent, child);
+                &self->tables->edges, x->left, x->right, parent, child, NULL, 0);
             if (ret < 0) {
                 goto out;
             }
@@ -6261,9 +6449,9 @@ tsk_table_collection_load_indexes(tsk_table_collection_t *self)
     const char *removal_order = "indexes/edge_removal_order";
     read_table_col_t read_cols[] = {
         { insertion_order, (void **) &self->indexes.edge_insertion_order,
-            &self->edges.num_rows, 0, KAS_INT32 },
+            &self->edges.num_rows, 0, KAS_INT32, 0 },
         { removal_order, (void **) &self->indexes.edge_removal_order,
-            &self->edges.num_rows, 0, KAS_INT32 },
+            &self->edges.num_rows, 0, KAS_INT32, 0 },
     };
 
     has_insertion = kastore_containss(self->store, insertion_order);
@@ -6873,6 +7061,10 @@ tsk_squash_edges(tsk_edge_t *edges, tsk_size_t num_edges, tsk_size_t *num_output
     j = 0;
     l = 0;
     for (k = 1; k < num_edges; k++) {
+        if (edges[k - 1].metadata_length > 0) {
+            ret = TSK_ERR_CANT_PROCESS_EDGES_WITH_METADATA;
+            goto out;
+        }
 
         /* Check for overlapping edges. */
         if (edges[k - 1].parent == edges[k].parent